requirements.txt
Flask
Werkzeug
chardet
numpy
pandas
polyfuzz
sentence-transformers
faiss-cpu